Add core directory to pipeline
