+++ apps/pipeline-asset-sg/src/core/config.ts
@@ -1,0 +1,39 @@
+export interface PipelineConfig {
+  allowedWorkgroupIds: number[];
+  source: {
+    connectionString: string;
+  };
+  destination: {
+    connectionString: string;
+  };
+}
+
+function getEnvOrThrow(key: string): string {
+  const value = process.env[key];
+  if (!value) {
+    throw new Error(`Missing environment variable: ${key}`);
+  }
+  return value;
+}
+
+export function getConfig(): PipelineConfig {
+  return {
+    allowedWorkgroupIds: getEnvOrThrow('ALLOWED_WORKGROUP_IDS').split(',').map(Number),
+    source: {
+      connectionString: getEnvOrThrow('SOURCE_CONNECTION_STRING'),
+    },
+    destination: {
+      connectionString: getEnvOrThrow('DESTINATION_CONNECTION_STRING'),
+    },
+  };
+}
+
+export function prismaConfig(connectionString: string) {
+  return {
+    datasources: {
+      db: {
+        url: connectionString,
+      },
+    },
+  };
+}
