+++ apps/pipeline-asset-sg/src/core/export-to-view.service.ts
