--- apps/pipeline-asset-sg/src/utils.ts
@@ -1,7 +1,0 @@
-export const prismaConfig = (connectionString: string) => ({
-  datasources: {
-    db: {
-      url: connectionString,
-    },
-  },
-});
